# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ApareText API",
    "description": "API REST para gesti\u00f3n de snippets",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/snippets": {
      "get": {
        "summary": "List Snippets",
        "description": "Listar todos los snippets.",
        "operationId": "list_snippets_api_snippets_get",
        "parameters": [
          {
            "name": "enabled_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filtrar solo snippets habilitados",
              "default": false,
              "title": "Enabled Only"
            },
            "description": "Filtrar solo snippets habilitados"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por tags (CSV)",
              "title": "Tags"
            },
            "description": "Filtrar por tags (CSV)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Snippet"
                  },
                  "title": "Response List Snippets Api Snippets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Snippet",
        "description": "Crear nuevo snippet.",
        "operationId": "create_snippet_api_snippets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnippetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snippet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/snippets/{snippet_id}": {
      "get": {
        "summary": "Get Snippet",
        "description": "Obtener snippet por ID.",
        "operationId": "get_snippet_api_snippets__snippet_id__get",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snippet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snippet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Snippet",
        "description": "Actualizar snippet existente.",
        "operationId": "update_snippet_api_snippets__snippet_id__put",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snippet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnippetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snippet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Snippet",
        "description": "Eliminar snippet.",
        "operationId": "delete_snippet_api_snippets__snippet_id__delete",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snippet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/snippets/search/{query}": {
      "get": {
        "summary": "Search Snippets",
        "description": "Buscar snippets.",
        "operationId": "search_snippets_api_snippets_search__query__get",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por tags (CSV)",
              "title": "Tags"
            },
            "description": "Filtrar por tags (CSV)"
          },
          {
            "name": "scope_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por scope",
              "title": "Scope Type"
            },
            "description": "Filtrar por scope"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Snippet"
                  },
                  "title": "Response Search Snippets Api Snippets Search  Query  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/snippets/expand": {
      "post": {
        "summary": "Expand Snippet",
        "description": "Expandir snippet con variables.",
        "operationId": "expand_snippet_api_snippets_expand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnippetExpand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnippetExpandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/abbreviations/{abbreviation}": {
      "get": {
        "summary": "Get By Abbreviation",
        "description": "Obtener snippet por abreviatura.",
        "operationId": "get_by_abbreviation_api_abbreviations__abbreviation__get",
        "parameters": [
          {
            "name": "abbreviation",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Abbreviation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Snippet"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get By Abbreviation Api Abbreviations  Abbreviation  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Get Stats",
        "description": "Obtener estad\u00edsticas de uso.",
        "operationId": "get_stats_api_stats_get",
        "parameters": [
          {
            "name": "snippet_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snippet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/import": {
      "post": {
        "summary": "Import Snippets",
        "description": "Importar snippets desde JSON.",
        "operationId": "import_snippets_api_import_post",
        "parameters": [
          {
            "name": "replace",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Reemplazar snippets existentes",
              "default": false,
              "title": "Replace"
            },
            "description": "Reemplazar snippets existentes"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validate-template": {
      "post": {
        "summary": "Validate Template",
        "description": "Validar sintaxis de template.",
        "operationId": "validate_template_api_validate_template_post",
        "parameters": [
          {
            "name": "template",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ScopeType": {
        "type": "string",
        "enum": [
          "global",
          "apps",
          "domains"
        ],
        "title": "ScopeType",
        "description": "Tipos de scope para snippets."
      },
      "Snippet": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "abbreviation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          },
          "snippet_type": {
            "$ref": "#/components/schemas/SnippetType",
            "default": "text"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "content_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Html"
          },
          "is_rich": {
            "type": "boolean",
            "title": "Is Rich",
            "default": false
          },
          "image_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Data"
          },
          "scope_type": {
            "$ref": "#/components/schemas/ScopeType",
            "default": "global"
          },
          "scope_values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scope Values"
          },
          "caret_marker": {
            "type": "string",
            "title": "Caret Marker",
            "default": "{{|}}"
          },
          "variables": {
            "items": {
              "$ref": "#/components/schemas/SnippetVariable"
            },
            "type": "array",
            "title": "Variables"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Snippet",
        "description": "Snippet completo con validaci\u00f3n."
      },
...
      "SnippetExpand": {
        "properties": {
          "snippet_id": {
            "type": "string",
            "title": "Snippet Id"
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables",
            "default": {}
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "target_app": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target App"
          },
          "target_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Domain"
          }
        },
        "type": "object",
        "required": [
          "snippet_id"
        ],
        "title": "SnippetExpand",
        "description": "Request para expandir snippet."
      },
...
      "SnippetType": {
        "type": "string",
        "enum": [
          "text",
          "image"
        ],
        "title": "SnippetType",
        "description": "Tipos de snippets."
      },
...
    }
  }
}